CurrencyAdapter refactor and used in frontend

+++ src/main/java/com/example/juiceFactory2_0/adapter/Currency.java
@@ -1,5 +1,7 @@
 package com.example.juiceFactory2_0.adapter;
 
+import java.math.BigDecimal;
+
 public interface Currency {
-    double getAmount();
+    BigDecimal getAmount();
 }
